% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_park_factor.R
\name{base_park_factor}
\alias{base_park_factor}
\title{Calculate the 5-Year Base Park Factor for an NCAA Baseball Team's Home Stadium}
\usage{
base_park_factor(team_id)
}
\arguments{
\item{team_id}{Numeric or character. The unique identifier for the team whose park factor is being calculated.}
}
\value{
A single numeric value representing the 5-year base park factor for
the team's home stadium
}
\description{
Calculate the 5-Year Base Park Factor for an NCAA Baseball Team's Home Stadium
}
\examples{
base_park_factor(124)

 @export base_park_factor

}
