% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_guts.R
\name{load_guts}
\alias{load_guts}
\title{Load and Prepare Guts Data for NCAA Baseball Statistics}
\usage{
load_guts()
}
\value{
A data frame of the guts statistics for the 2023 season, with the season column removed and
the \code{wOBA} column renamed to \code{lgwOBA}. The data includes various metrics that are essential for in-depth baseball analysis.
}
\description{
This function downloads and processes a CSV file containing "guts" statistics for NCAA baseball.
"Guts" stats provide a comprehensive overview of various baseball metrics.
The data is specifically filtered for the 2023 season, and the \code{wOBA} column is renamed to \code{lgwOBA} for clarity.
}
\examples{
guts_data <- load_guts()

}
