% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_teams.R
\name{load_teams}
\alias{load_teams}
\title{Load NCAA Baseball Teams Data for 2024}
\usage{
load_teams()
}
\value{
A data frame containing information about NCAA baseball teams for the year 2024. The data frame includes
various details about each team, tailored to the 2024 season.
}
\description{
This function retrieves data on NCAA baseball teams for the year 2024. It makes use of the \code{baseballr} package
to fetch the teams' information and then filters the dataset to include only teams for the specified year.
}
\examples{
teams_2024 <- load_teams()

}
