% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/park_factor.R
\name{park_factor}
\alias{park_factor}
\title{Calculate Adjusted Park Factor for an NCAA Baseball Team}
\usage{
park_factor(team_id)
}
\arguments{
\item{team_id}{Numeric or character. The unique identifier for the team whose park factor is being calculated.}
}
\value{
A single numeric value representing the adjusted park factor for the team, accounting for the proportion of
home games and the deviation of the team's basic park factor from 1. The park factor is adjusted to mitigate the influence
of extreme values, aiming to provide a balanced metric that reflects the impact of the team's stadium on game results.
}
\description{
This function calculates the adjusted park factor for an NCAA baseball team for the year 2024, taking into account the
team's schedule, home game frequency, and the basic park factor calculated by the \code{base_park_factor} function. The adjusted
park factor provides a more nuanced understanding of how a team's home stadium influences game outcomes compared to league averages.
}
\examples{
adjusted_pf <- park_factor(124)

}
