% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_id_to_name.R, R/team_name_to_id.R
\name{team_name}
\alias{team_name}
\title{Retrieve the School Name of an NCAA Baseball Team by Team ID}
\usage{
team_name(team_name)

team_name(team_name)
}
\arguments{
\item{team_id}{Numeric or character. The unique identifier for the team whose name is being retrieved.}
}
\value{
A character string representing the school name of the specified team. If no match is found, the function
returns \code{NULL}.

A character string representing the school name of the specified team. If no match is found, the function
returns \code{NULL}.
}
\description{
This function queries a dataset of NCAA baseball teams to find the school name associated with a specific team ID.
It relies on the \code{load_teams} function to fetch the current list of teams, filtering this list to match the specified
team ID and extracting the school name.

This function queries a dataset of NCAA baseball teams to find the school name associated with a specific team ID.
It relies on the \code{load_teams} function to fetch the current list of teams, filtering this list to match the specified
team ID and extracting the school name.
}
\examples{
# Assuming '12345' is a valid team_id
school_name <- team_name(12345)

# Assuming '12345' is a valid team_id
school_name <- team_name(12345)

}
