% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_name_to_id.R
\name{team_name_to_id}
\alias{team_name_to_id}
\title{Retrieve the School ID of an NCAA Baseball Team by Team Name}
\usage{
team_name_to_id(name)
}
\arguments{
\item{name}{Character. The name of the team whose id is being retreived.}
}
\value{
A number representing the school id of the specified team. If no match is found, the function
returns \code{NULL}.
}
\description{
This function queries a dataset of NCAA baseball teams to find the school name associated with a specific team ID.
It relies on the \code{load_teams} function to fetch the current list of teams, filtering this list to match the specified
team ID and extracting the school name.
}
\examples{
# Assuming '12345' is a valid team_id
school_name <- team_name(12345)

}
