% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_stats.R
\name{team_stats}
\alias{team_stats}
\title{Retrieve and Process NCAA Baseball Team Statistics}
\usage{
team_stats(team_id, type = "batting")
}
\arguments{
\item{team_id}{Numeric or character. The unique identifier for the team whose statistics are being retrieved.}

\item{type}{Character. Specifies the type of statistics to retrieve: 'batting' or 'pitching'.
Defaults to 'batting'.}
}
\value{
A data frame containing the processed statistics for the team's players. For batting,
the statistics include games played, plate appearances, home runs, runs, RBIs, stolen bases,
walk and strikeout percentages, ISO, BABIP, batting average, on-base percentage, slugging percentage,
wOBA, and wRC+. For pitching, the statistics include games played, games started, innings pitched,
strikeouts per 9 innings, walks per 9 innings, home runs per 9 innings, BABIP, ERA, and FIP.
}
\description{
This function fetches and processes team statistics for an NCAA baseball team for the year 2024,
based on the team ID and the type of statistics (either 'batting' or 'pitching'). The function
integrates park factor adjustments for batting statistics and offers detailed metrics such as wOBA,
wRC+, and others for batting, as well as FIP, K/9, and BABIP for pitching.
}
\examples{
batting_stats <- team_stats(124, "batting")
pitching_stats <- team_stats(124, "pitching")

}
